FP128 optimization: Add Lifetime Markers for FP128ptr
In details:
As for LifetimeStart, normally, we will create LifetimeStart before Store. But if the a fp128 value is defined inside a loop and used outside the loop, we will hoist LifetimeStart outside the loop safely.
As for LifetimeEnd, when there are no more uses of a value on a code path, a lifetime.end marker can be inserted after the final use.

Change-Id: I487dcfbefbb52f16d3ec4251ab6c8081c16ab43e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/237546
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>