[Driver][DPCPP] Disable options for DPC++
For DPC++, there are certain options we want to restrict and not have
output in the help, as they are confusing to users or we have explicitly
disabled support.  OpenMP enabling and cuda options are part of this.

Also add a way to just suppress an option from the help instead of fully
disabling.

Change-Id: I2bfe12a76dca063f907ea442f876bd24018a9be6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/237974
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>