CMPLRLLVM-18347 - Problem with mkl performance libs with -fsycl
After a recent merge, we lost some functionality with the MKL performance
libraries when -fsycl is present.  Was not caught by testing as it looks like
there were some spelling errors in the test...

Did a bit of refactoring to clean up the code to make sure that the offload
static lib transform is done only once.

Change-Id: Id63ed47963c27e623ff6c43c00642adf917a846e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238012
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>