[MapIntrinToIml] Enable widening of multi-return value (e.g. sincos and divrem) SVML calls.
This patch is part of the effort to rectify internal signatures of SVML
functions.

Before this patch, MapIntrinToIml expects sincos calls in the form below:

void @__svml_sincosf4(<4 x float>, <4 x float *>, <4 x float *>)

And will lower it to calls returning vectors with 2*VF and a series of stores
to the pointers.

This patch requires sincos functions to be called like:

{ <4 x float>, <4 x float> } @__svml_sincosf4(<4 x float>)

And lowers it to calls returning structures of vectors after performing
legalization. This matches the implementation of the SVML library.

Currently only handling of widening is implemented. To fully enable
legalization of sincos/divrem SVML calls, splitting also needs to be enabled,
this will be done in the next patch.

JIRA: CMPLRLLVM-11328, CMPLRLLVM-9746
Change-Id: I3b99497ffe87829ba76b558aa05ca90b0c6c2fb1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238076
Reviewed-by: Tianqing Wang <tianqing.wang@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>