[NFC][VPlan] Check output right after Plain CFG creation in some tests
Some tests (most are HIR-based in this commit) are targeting the Plain CFG
construction but for some reason CHECK the output after some inner loop
massaging is executed in the pipeline. While the dumps are the same currently,
the tests are unnecessary exposed to fails due to changes in the pipe line not
related to their particular purpose.

Fix them by modifying the RUN lines to output dumps right when the targeted
thing happens.

Change-Id: I22fda83a8e2838eb9b08af86d0a14b5decc35063
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238224
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>