[MapIntrinToIml] Compile-time selection of SVML function variants for different ISAs.
When migrated to xmain, libiml_attr's target detection and related
features were disabled since they rely on ICC specifics. This patch
uses information in Subtarget to determine ISA for SVML variant selection.
This also means libiml_attr was only using mappings for IA32 on xmain.
With this patch, this is also fixed and libiml_attr is now aware of the
target architecture.

Change-Id: I71b5fa4be86a41b8ff47488a26ba06df4ae62539
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/304354
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Tianqing Wang <tianqing.wang@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>