Fix segfault of ompo_kernels_small_gpuCpp tests
D74837 [1] changed the way how global state of libomptarget lirary is
stored within it from global variables to dynamically allocated variables
with functions explicitly marked as global constructor/destructor and
priority.

The problem is that aforementioned attributes are not standard and not
supported by MSVC.

This patch added 'DllMain' function which is intended to do the same job.
Code is based on the example from officla documentation [2]

There are some other tricks mentioned on stackoverlow [3], but they seem
to be too hacky.

[1]: https://reviews.llvm.org/D74837
[2]: https://docs.microsoft.com/en-us/cpp/build/run-time-library-behavior?view=vs-2019
[3]: https://stackoverflow.com/questions/1113409/attribute-constructor-equivalent-in-vc

Change-Id: I3f91fdbe285caf984fbabfe1a2a23f8386a52475
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238619
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: Battini, Prasad V <prasad.v.battini@intel.com>