CMPLRLLVM-11402: Part 3 of multiversioning and inlining for tiling. This transforms the IR by introducing the multiversioning test and transforming the high performance version by marking the tile choices for inlining. The tile choices are later inlined by the inliner.
There will be one more change set after this one. I just noticed that down-
stream optimizations do not completely eliminate all of the redundant global
variable references and branches that they could.  So, I will strengthen the
simplification of the conditionals to take care of that. This change set is
big enough that it is better to do that in a follow up change set.

Change-Id: I0f7c6de19a793d730d644db12db94a1f742e73ba
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238851
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>