[SLP] Prioritize vectorization of pairs which are operands of same cmp.
For cmp instructions which contribute to a reduction it is important to
pair into a vector operands of the same instruction (the reduction,
comparisons and their operands have all to be defined in same basic block).
To achieve that we (1) bail out vectorizing a cmp instruction if any of it
operands defined in another basic block and (2) once a reduction is matched
but failed to vectorize as a whole we make attempt to vectorize compares
right there as a priority.

Fixes CMPLRLLVM-11933.

Change-Id: I894ae32d8ac112c63ac34315ebc4e3b38f60e69d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238878
Reviewed-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>
Tested-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>