[VPlan][NFC]: Adding routine name to CM dumps.
Routine name altogether with loop ID is added into output of CM dumps.
That makes it easier locating CM dump in output of flto compilation.

CM dump routine is now invoked after predicate pass which makes the numbers in
the dump to be closer to CM numbers that are used during VF selection.

Change-Id: Id533fb40a1e849947e100bdb3492908c9d2ba8fd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/239053
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>