CMPLRLLVM-18421 - Re-emit only used VLA size expressions.
Prior to this change, all VLA expressions in the VLASizeMap were being
re-emitted for each region.  That strategy failed when VLAs are declared
in multiple scopes.

Now only re-emit size expressions that are used in the captures on the
region. Also move the logic up to CodeGenFunction since it needs to occur
once per directive to correctly handle combined directives with multiple
Outliner objects. Also handle nested OpenMP directives with VLAs.

Change-Id: I8dc0a86d8f5b92af37fddc270b98d18b572e0206
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/239191
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: Rice, Michael P <michael.p.rice@intel.com>