[NFCI][VPlan] Don't make wrong update to LoopInfo on the fly
The code to perform on-the-fly updates to LoopInfo seems to remain from the
times when innermost-only vectorization was supported, it is obviously wrong
now.

We don't claim the analysis to be preserved by the VPlanDriver and the wrong
information stored inside the analysis doesn't cause any known issues, so I
suppose we don't try to [mis-]use it in any way, hence NFCI tag.

Change-Id: I25ddec44c0a5fc6928c7ec0855b1993f48b657d6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/239398
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>