[LoopOpt] Improve the correctness of alias analysis queries
LoopOpt queries alias analysis to try to eliminate entire dependence
arcs. For example, in

for(j)
  B[j] = A[j] + 1;

...if AA can say that B and A noalias, then no dependence arc is needed.
However, this is not always safe. AA is only asserting that two
contemporary SSA values do not alias. I.e., it's reasoning about two
pointers assumed to result from the same dynamic path and state. As a
result, the AA query is only useful where we can prove that the pointers
are loop invariant. For example, below, a query on {B, A} is only useful
in the context of the 'j' loop.

for(i) {
  for(j)
    B[j] = A[j] + 1;
  swap(B, A);
}

Before this change, we use such a query in the context of the entire
loop nest and incorrectly determine that there is no dependence at all.

This change ensures that we do not use such a query to eliminate the
entire dependence arc between the two references by first checking that
the pointers queried are loop invariant. If the pointers are not
both invariant but the base pointers are, then a less precise query with
stronger semantic guarantees about its result will be made.

If a base pointer is not invariant then the AliasAnalysis "alias"
interface does not make a strong enough assertion to be useful, as
described above. In this case, DDTests uses a new "loopCarriedAlias"
interface, which asserts that the two values don't alias even across
different execution points in the program. HIRSymbaseAssignment also
uses this interface indirectly via a new "HybridAliasSetTracker".

Change-Id: If2226734ef63b9ac8782fa2dab68749da145a5f0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242463
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>