Fix a number of tests based on 2 changes. fno-common & -fsycl
First, -fno-common became default for c mode, so all tests that had a
'common' in their global print line has it removed. None of the tests
seemed to be depending on the existence of common.

Second, 'SYCL' mode now requires BOTH -fsycl AND -fsycl-is-XXX. This
change made a bunch of tests that depended on being in SYCL mode no
longer work.

Change-Id: Ia88de76f766da2ce6371dcf5bc550474c34bc2ea
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/239586
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>