[VPlan] Updates eraseInstruction() to do the following:
1. unlinks VPInstruction from VPBasicBlock's Instructions list.
2. adds VPInstruction in UnlinkedVPInsns vector which holds the VPInstructions
   that need to be deleted by VPlan's destructor.

This change is needed because the following problem might occur. If we erase a
VPInstruction without removing it from all the data structures that is used,
then its pointer remains in the data structures. If a new VPInstruction has the
same pointer as the one that was erased, then the data structures have wrong
values. Hence, it is better to completely erase all VPInstructions at the end
of VPlan.

Change-Id: I25603c0acac0e6c938e8e122bbf66ead9c7772f5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/239688
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>