[X86] Fix FsFLD0SS/FsFLD0SD bug of global FMA. When do global FMA optimization there is below machine IR transform.
  %6950:fr64x = FsFLD0SD
GENERATE NEW INSTRUCTION:
  %6949:fr64x = nofpexcept VFMSUB213SDZr %3670:vr128x(tied-def 0),
                %%334:vr128x, %3724:fr64x, implicit $mxcsr
GENERATE NEW INSTRUCTION:
   %3747:fr64x = nofpexcept VFNMADD213SDZr killed
                 %%6949:fr64x(tied-def 0), %3739:fr64x, %6950:fr64x,
                 implicit %$mxcsr
DELETE the MI (it is replaced):
   %3746:fr64x = nnan ninf nsz arcp contract afn reassoc
                 %nofpexcept VSUBSDZrr %3724:fr64x, %3723:fr64x,
                 %implicit $mxcsr
DELETE the MI (it is replaced):
   %3723:fr64x = nnan ninf nsz arcp contract afn
                 %reassoc nofpexcept VMULSDZrr %3670:vr128x,
                 %%334:vr128x, implicit $mxcsr
DELETE the MI (it is replaced):
                 %3747:fr64x = nnan ninf nsz arcp contract
                 %afn reassoc nofpexcept VMULSDZrr
                 %%3739:fr64x, %3746:fr64x, implicit $mxcsr
In high register pressure, %6950 is allocated xmm24. However FsFLD0SD
is not AVX512 instruction, and the register class for FsFLD0SD is fr64,
so when do instruction encoding, the bit 4 for register number 24 is
ignored and xmm24 become xmm8. So error happens.

The fix is to generate the AVX512 version of FsFLD0SD and FsFLD0SS,
so that it can accept higher register number (xmm16 ~ xmm31).

Change-Id: If4e254132458a2153a1bfa1c520ce0eed547633d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/239965
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: Luo, Yuanke <yuanke.luo@intel.com>