[VPlan] Changes to avoid decomposing invariant canon expressions during HIR decomposition
Changes here avoid decomposing canon expressions in a DDREF that are invariant
during VPlan construction of HIR loops. Consider the following loop being
considered for vectorization.

  for (i1 = 0; i1 < 100; i1++) {
     arr[n1 * n2][i1] = 111;
     farr[n1 + n2 + i1] += farr2[n1 * n2 + i1];
   }

The canon expression [n1 * n2] in the reference to arr is invariant.
Similarly [n1 + n2] and [n1 * n2] are invariant in the canon expressions
that reference farr/farr2 once we ignore the loop IV. The changes here
generate unique external defs for [n1 * n2] and [n1 + n2]. The external defs
corresponding to [n1 * n2] and [n1 + n2] are added to the IV appropriately
during decomposition when generating the accesses to farr/farr2.

Changes to avoid decomposing invariant DDREFs and invariant Blobs will be
uploaded as separate patches.

Change-Id: I43198fbe79a1437c3153151c8d252dca31d36e9d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/240330
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>