[Driver][SYCL][FPGA] Adjust device and AOCX link order for FPGA (#1389)
This is a cherry pick of:
https://github.com/intel/llvm/commit/ba404be42c028934c3e33da1239e526975d51ef6

When compiling with -fintelfpga, a user can provide an -fsycl-link=image
created aocx based binary/archive.  When linked in, this archive should get
precedence over any other device object that is extracted from object or
generated from source to ensure the proper binary is grabbed at runtime.

Change-Id: Ie56172cb0f4b42b040b70f94cfefa050588c18e8
Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/240353
Reviewed-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>