CMPLRLLVM-10481: Add LINEAR:IV clause and increment final expressions for simd counter variables.
This adds two new expressions to the AST: the linear step and
an increment of the linear counter.

This adds new implicit clause kinds: ICK_linear_private and
ICK_linear_lastprivate.  This aids in the addition of linear to the
SIMD directive and private/lastprivate to the outer loop.  They
are also used to decide whether the increment is needed.  When the
loop counter is defined outside the loop linear_lastprivate is used.

LINEAR:IV is added only for scalar and pointer variables.
The question is open on what to do about C++ iterator variables.

Change-Id: Iae27702de04188731433bb1001d70de317df089d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/240357
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: Rice, Michael P <michael.p.rice@intel.com>