Fix crash when transforming OpenCL channel to FPGA pipe
The InstCombinePass introduced dangling Constant Users of channels
which are PtrToInt operators. These Constants are out of scope of
ChannelPipeTransformationPass, which leads to cast failure when it
performs transformations.

The solution is to process them rather than remove them because we
can not assert that they will not be used later, even though they
are dead constant users dangling off of the channels.

Change-Id: Ib20f677ce12789d43e33b1fc86e4dcd47013eb64
Signed-off-by: Mikhail Lychkov <mikhail.lychkov@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/240479
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>