[SYCL][FPGA] Fix sporadic failure of SYCLIOPipes.ll
Summary:
llvm::StringRef -> std::string

It is recommended to use std::string when the value is computed,
or there are other lifetime issues and use StringRef otherwise.

Details:
Here comes a difference between I/O pipes/channels in OpenCL and in
SYCL/DPCPP. In OpenCL a user manually specifies name of I/O pipe/channel
with const char * string, which will be encoded in LLVM IR as string
metadata attached to pipe/channel GV later, in SYCL/DPCPP I/O pipes are
having integer metadata attached to a pipe GV.

Since this metadata is used by FPGA emulator to open the appropriate binary
file and therefore the SYCL/DPCPP version of the metadata is needed to be
mutated to string in our backend. In this case it's better to use std::string
instead of llvm::StringRef.

Change-Id: I7ee1bc31b76c173e869c32ecbe98ef865da5cc39
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/240455
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>