Correctly handle _may_i_use_cpu_feature with a zero parameter.
I unfortunately missed validating situations where the parameter to the
_may_i_use_cpu_feature (and _ext version) were zero was passed as the
request. This patch makes us properly handle that situation by resulting
in a 'true' expression.

While we DO have the ability to emit an error in the case where someone
attempts to ask if they can use 'no' features, it seems that this would
be easy enough to cause with a macro or template instantiation, so I've
instead chosen to just give a valid answer.

Change-Id: I9d198c9187d78dfdd1c16eef66add3a1b207acfb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/240492
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>