[VPlan][2/2] Create vplan-function-vectorizer pass
Currently it is expected to be used for LIT testing of VPlan's general
transformations that don't have specific assumptions about being done on
LLVM/HIR or loop/region vectorization.

In future, we might extend it into a full-blown transformation for kernel
vectorization and/or OMP function vectorization.

For now, vectorize the whole function body and assume all the parameters are
uniform and that the function has a single dedicated "ret void" exit.

Change-Id: I3e468dca1ba1a7c2c8f98a8085dbec0f25f3928f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/241999
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>