[VPlan][1/2] Refactor LLVM IR CFG -> VPlan CFG utility into a separate class
This is needed to start implementing a function vectorization pipeline. The main
functionality is moved into a base class using CRTP pattern.

In the next commit(s) I will use that same base class to implement CFG builder
for the basic whole function vectorization that is going to simplify many our
LIT tests.

Change-Id: Iead3c9b96874715a3ecfc3fb658c1200ede63d10
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/241998
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>