[SYCL] Enable non-read-write memory object mapping in scheduler (#1390)
Previously all memory objects were mapped with read-write access
whenever they were requested on host so that they were fully
accessible on host until their unmapping. This patch changes this
behaviour: now memory objects are mapped with just the required access
mode. If an incompatible access mode is requested on host afterwards,
the object is remapped.

Signed-off-by: Sergey Semenov <sergey.semenov@intel.com>

This is a cherry-pick of intel/llvm PR #1390:
https://github.com/intel/llvm/pull/1390

This fixes the following issue:
https://jira.devtools.intel.com/browse/CMPLRLLVM-15482

Change-Id: Ie4ee90ae7327c0dfa61d260b381c84c87d5854da
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/240976
Reviewed-by: Semenov, Sergey <sergey.semenov@intel.com>
Tested-by: Semenov, Sergey <sergey.semenov@intel.com>