[DPCPP][HOST] Port DPCPPKernelVecClone pass over from OpenCL
This is a part of effort to enable DPCPP kernels on the HOST through
WGLoopCreator/Barrier approach.

The goal is to enable a subset of benchmarks by doing this, not provide
product quality from day #1.
This port does not aim to port complete functionality from OpenCL to
DPCPP, only core functionality that is needed achieve the above.
So unless problems detected in review are critical to this goal, I'd
rather file JIRAs and address them later.

Change-Id: I068af2e4128ae66d53c585c132d08b71e6b0b6ed
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/241743
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>