[MapIntrinToIml] Enable splitting of multi-return value (e.g. sincos and divrem) SVML calls.
This patch is part of the effort to rectify internal signatures of SVML
functions. As widening has been implemented in a previous patch, this patch
completes multi-return function support in MapIntrinToIml component.

Previously, MapIntrinToIml expects sincos calls with pointers as output
parameters:

void @__svml_sincosf4(<4 x float>, <4 x float *>, <4 x float *>)

And will lower it to calls returning vectors with 2*VF and a series of stores
to the pointers.

This patch requires sincos functions to be called like:

{ <4 x float>, <4 x float> } @__svml_sincosf4(<4 x float>)

The legalized calls also return structures of vectors. This matches the
implementation of the SVML library better, and greatly simplifies the code.

JIRA: CMPLRLLVM-11328, CMPLRLLVM-9746
Change-Id: I3fbe40b34298464595afa616d60454b686f7daaf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/241496
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>