[SYCL] Share PFWG lambda object through shared memory (#1455)
In the current implementation private address of the PFWG lambda object
is shared by leader work item through local memory to other work items.
This is not correct. That is why perform the copy of the PFWG lambda
object to shared memory and make work items work with address of the
object in shared memory. I.e. this case should be handled in the
similar way as for byval parameters.

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>
Change-Id: I3ca2d52b4a2b063932ad048a1785d57f9469c8e9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/241805
Reviewed-by: Bobrovsky, Konstantin S <konstantin.s.bobrovsky@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>