[Loopopt] Changes to fix VLS lit test fails caused by ShuffleVectorInst changes.
There was a recent change to store the mask value as a SmallVector instead of
as an operand in ShuffleVectorInst. This means ShuffleVectorInst now only has
two operands. However, in HIR we continue using three operands for a
ShuffleVectorInst representation. When we call ShuffleVectorInst constructor
in HIRCG, the constructor takes care of storing the mask in the SmallVector
from the mask Value argument.

Change-Id: I27fac6ae7bbedfab667b1c715ae94ab24af03de9
Jira: CMPLRLLVM-18881
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/241901
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>