[DTRANS] CMPLRLLVM-18898: Fix for delete fields in DTrans
Delete fields can be applied to a structure with nested structures,
but does not necessarily need to be applied to its inner structures.
In this case there is no information in the types map for the
inner structures. If there is no information for the inner
structures then the index the GEP will be the same.

This change set prevents the check in the types map if the
information for the inner structure isn't available and uses the
index that is in the GEP. This prevents a segmentation fault.

Change-Id: Ie900f176a49df404991895b3bbb128c9c2ce433b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/241991
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>