[DPCPP][HOST] Port DPCPPKernelWGLoopCreator from OpenCL
This is a part of effort to enable DPCPP kernels on the HOST through
WGLoopCreator/Barrier approach.

The goal is to enable a subset of benchmarks by doing this, not provide
product quality from day #1.
This port does not aim to port complete functionality from OpenCL to
DPCPP, only core functionality that is needed achieve the above.
So unless problems detected in review are critical to this goal, I'd
rather file JIRAs and address them later.

Change-Id: I12f284f052c80aeacdaf341d68a1e7379ba93b7e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242101
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>