[DTrans - OpaquePointers] Add command line option to filter trace messages
When debugging the pointer type analyzer on applications, printing the IR
statements and their dependencies can be extremely verbose and slow due
to the way printing IR instructions needs to compute the variable names on
each print call. This enables a subset of the trace to be printed by
adding -dtrans-pta-filter-funcs option to take a list of functions to limit
the verbose trace messages to specific functions.

Change-Id: Ib194f2bae0890c4fb875ab7a3dfa38e8735ab691
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242206
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>