CMPLRLLVM-18525: Fix for support GNU-LTO and archives
The community made some changes on the naming for archived
bitcode files (https://reviews.llvm.org/D75422). This new
naming can cause a conflict when identifying if file
inside a thin archive is a GNU archive or a Bitcode.

This changeset checks if the archived file inside a thin
archive is a bitcode file, and if so, does not try to create
a binary object out of it.

Change-Id: If2ec0c0edcac0fc3ed56b678b50be60ceb2bce1c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242403
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>