Tighten up test for constant propagation into outlined loops
This test was introduced alongside IPO/PassManagerBuilder changes to do
IP constant propagation with VPO paropt at O3+. The test runs "opt -O3"
to test this, but this results in a different pass schedule which fails
to do the propagation in question. This went unnoticed because the test
falsely passes: an inlined copy of the outlined function sees the
propagation, but the outlined function does not. The test's CHECK passes
based on the inlined propagation.

To fix the test's checking, this change adds a CHECK-LABEL to ensure
we're checking the outlined function and not an inlined copy. To make
the interprocedural propagation actually happen, it begins with IR
straight out of the FE and additionally specifies "-paropt=31" to opt.

Change-Id: Ifdcfb754f9d87892c4bbdafae286ad0d6f3a8d08
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238375
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>