[JIT] Switch from MCJIT to LLJIT for module compile
LLJIT (ORCv2) serves as a replacement for MCJIT. ORC supports eager/lazy
compilation and concurrent compilation which could be of interest to us.
Since we are already using LLJIT to reload program binary, it is better
to use LLJIT for compilation as well.
This patch also improves module ownership management.

ExecutionEngine codes are still preserved for LLDJIT.

Change-Id: Ib3824c0d4211bd96c0ea7f55dfc20ce2a99ecfb7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/227538
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>