[Driver][SYCL] Move include/sycl header before other system header locations (#1492)
This is a cherry pick of:
https://github.com/intel/llvm/commit/2b88867304947730ffc02b2c7a70abaa753ac515

The location in which the include/sycl headers are added are too late in
the ordering sequence.  This is causing the possibility of the wrong version
of the OpenCL headers to be picked up (i.e. from /usr/include)

Improve the ordering by putting the dirs much earlier in the command line
instead of relying on the ToolChain addition.

Change-Id: Ic658edd74c67e9793eb6f85bfad7b30000e7076a
Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242620
Reviewed-by: Chilamantula, Harini <harini.chilamantula@intel.com>
Reviewed-by: Thatishetti, Meghana <meghana.thatishetti@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>