[DPCPP][Host] Add WGLoopCreator, Pre/Post Vec passes into IPO pipeline
This is a part of effort to enable DPCPP kernels on the HOST through
WGLoopCreator/Barrier approach.

The goal is to enable a subset of benchmarks by doing this, not provide
product quality from day #1.
This port does not aim to port complete functionality from OpenCL to
DPCPP, only core functionality that is needed achieve the above.
So unless problems detected in review are critical to this goal, I'd
rather file JIRAs and address them later.

Change-Id: I1f2a1bf68fa6999bf9b8bca00963b7e71454c66d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242102
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>