[AA] Add a new "loopCarriedAlias" interface to alias analysis
This is intended to be useful to dependence analyses. The normal interface can
only assert no alias between two contemporary SSA values which exist at the
same point in the execution of the program.

For example, the two references below do not alias:

    for(int j=0; j<100; j++)
        A[j] = A[j+1];

This is because the normal AA interface answers the question: given footprints
x and y, is there any single dynamic path d in which x and y may overlap?

However, in dependence analyses we are trying to answer a different question:
is there any pair of dynamic paths (d, e) for which x in d overlaps with y in
e?

Answering "no" to the latter question is a stronger assertion; it implies a
"no" to the former question. A "no" answer to the former does not imply a "no"
answer to the latter question, however. (Consider the example above.)

Some existing alias analyses, such as TBAA and Andersens always answer
the latter question. This change allows AA implementations to claim that
they answer this stronger question, and provides an "loopCarriedAlias"
interface which can be resolved only by those implementations.
Initially, ScopedNoAlias, TBAA, and Andersens respond to
"loopCarriedAlias" identically to "alias".

Since no analyses use loopCarriedAlias yet, this change introduces new
interpreter lines to all AndersensAA LIT tests which use 'aa-eval',
adding the '-evaluate-loopcarried-alias' option to check that
loopCarriedAlias benefits from at least Andersens. Once analyses start
using it, better end-to-end testing will be added.

Note that due to concerns about performance degradation, the interface
only guarantees the stronger semantics described above if both locations
use "UnknownSize" rather than precise sizes. It is the responsibility of
the AA client to ensure that they specify "UnknownSize" when they wish
to use the results to reason about a dependence between multiple points
in execution.

Change-Id: Icf4d2c310c4df7e18790695e17ddc753139fab72
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242462
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>