[SVML] Add svml support for exp2 intrinsics.
This change adds the exp2 intrinsic to svml mappings and comments out exp_finite
tests in svml-calls.ll because we do not translate those to svml in xmain. A
separate change will be pushed to the community to properly place these tests in
svml-calls-finite.ll, which is already marked as XFAIL for xmain.

Change-Id: I2cb4f0c64fc2f9f690a1dfb13b22815383dc9a79
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242686
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>