[VPlan][CG] Update underlying LoopInfo when adding new BasicBlock.
As a workaround for a SCEV fail on inconsistent LoopInfo, restored the old
hacky behavior when during CG we insert all new BasicBlocks into the outermost
loop. The right fix is to implement a correct way to update function's
LoopInfo: create inner loops when needed and place new blocks correctly.

See CMPLRLLVM-19018.

Change-Id: I6c3b91acc8f1e6d770f2b42322144282720c4c30
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242752
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>