[NFC][VPlan] Reorder function in the IntelVPlanDriver.cpp
Place definitions of the templated functions before their actual uses. I have to
do that because I am going to add an llvm::Loop explicit specialization for
runStandardMode<Loop> and all the explicit specialization have to come before
their use.

This patch is submitted on its own to reduce the diff in a future change.

Change-Id: Idd809b7147289e4558f98a6d9de2bc20724d397d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242961
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>