[VPlan][2/2] Full LoopInfo/SCEV invalidation after processing each loop
The current on-the-fly update is incomplete (doesn't update inner loops)
and should not be kept as-is. The testcase added is crashing if the incomplete
on-they-fly update is removed. For some reason no tests were added when the hack
was re-introduced.

Doing the complete update isn't trivial and I'd prefer to delay it at least
until the CG is simplified to not do any implicit CFG generation. As such,
perform the full LoopInfo recalculation after each loop is processed.

Change-Id: I00f4daa8252e1be5e3480e76486744ea52a0ea2c
Jira-Id: CMPLRLLVM-19018
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242962
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>