[DTrans] [CMPLRLLVM-19128] DTransAnalysis was missing some byte-flattened GEP accesses
When the IsElementZeroAccess rule triggers when evaluating bitcasts to i8*,
due to a structure beginning with an i8, i8*, or ptr-to-ptr type, the
possibility that the result is actually being used as a byte-flattened GEP
was missed. Extend the analysis to store the original type to handle this case.
Also, fix an issue with the order of value evaluation caused by the call graph
information collection that prevented byte-flattened GEPs from being evaluated
because type information was incomplete.

Change-Id:I99a70710d9913045666efae6a4d7f7a45d40cebe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/243187
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>