[DPCPP][HOST] Host device using OpenMP and OpenCL transforms.
This is a library part of host device implementation
using OpenMP for threading and OpenCL passes.

In this implementation the compiler generates a loop
over work-items for parallel_for with nd-range and
applies varios transformation over it.

parallel_for with a regular range is implemented using
a similar approach but with a hardcoded work-group size.
This value could be explicitly set via
DPCPP_PERF_HOST_WG_SIZE0/1/2 macros(per each dimension).

Barrier support is not yet implemented.

Change-Id: I5c9f76618d210bb7e874e6b4a517ec7c22c7b9ad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/244845
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>