[VPlan] LIT test to demonstrate incorrect decomposition during VPlan construction of HIR loops.
The test case below is added to demonstrate incorrect decomposition of an add
HLInst whose LVAL is a memref, i.e., an HLInst like a[i] = a + b. The current
decomposition generates an add instruction with three operands whose third
operand is the LVAL address. The correct decomposition should generate an
add using the first two operands followed by a store to the third operand.

Currently we generate:
 vp1 = add a, b, &a[i]

We should generate:
 vp1 = add a, b
 store vp1, &a[i]

The test will be updated to check for correct decomposition when the fix
is in place in an upcoming commit.

Change-Id: Ia96b3da54387ded5cbaf90232fc14d3fdda480e0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/243154
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>