[VPlan] Fix incorrect decomposition of HLInsts such as a[i] = b + c
When decomposing an HLInst like a[i] = a + b during VPlan construction,
decomposition generates an add instruction with three operands where the
third operand is &a[i]. The correct decomposition should generate an add
using the first two operands followed by a store to the third operand.

Currently we generate:
 vp1 = add a, b, &a[i]

We should generate:
 vp1 = add a, b
 store vp1, &a[i]

The changes here fix decomposition to correctly handle such HLInsts.

Change-Id: I9ccd8c4df9a6ca238b00485f366bcdcec33f305a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/243317
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>