[DPCPP][HOST] Port of PhiCanonicalization pass over from OpenCL
This is a part of effort to enable DPCPP kernels on the HOST through
WGLoopCreator/Barrier approach.

The goal is to enable a subset of benchmarks by doing this, not provide
product quality from day #1.
This port does not aim to port complete functionality from OpenCL to
DPCPP, only core functionality that is needed achieve the above.
So unless problems detected in review are critical to this goal, I'd
rather file JIRAs and address them later.

The pass is a prerequisite of main Barrier pass.

Change-Id: I5aa929f90948c4f35a20d6ef36f8214ae22e8c37
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/243680
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>