Implement clang -opt-report support for OpenMP pragmas.
Currently openmp pragmas on the kernel side are either ignored or
implemented. But at the final IR there are no way of knowing what are the
pragmas that were applied. Adding this information to the report file
would make this information explicit to understand the performance of the
application. This patch will add report of the pragmas to the optimization
report's file (CMPLRLLVM-9975).

Change-Id: Ibe482e2e2e8aed1d73aa5bc558cde33cf9d4fa28
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/242347
Reviewed-by: Ammarguellat, Zahira <zahira.ammarguellat@intel.com>
Tested-by: Ammarguellat, Zahira <zahira.ammarguellat@intel.com>